[Cast Streaming] Implement sender session class
This patch adds a new SenderSession class along with unit testing for
this class.

Embedders can create an instance of this class in order to start a
streaming session with an already known receiver. The embedder is
expected to have a valid TLS connection already established with
the receiver.

Bug: b/162542753
Change-Id: Ia91df728d947e1fa20ed0b360f595f75d41bf11c
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2378540
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>