[Cast Channel] Add broadcast routing in VirtualConnectionRouter.
Replaces TODO comments with the implementation. :)

This change is needed for a soon-upcoming patch that will introduce a
basic Cast Agent (service for handling Cast V2 Application Control
messages) for a Cast Receiver.

The CastMessageHandler interface API contract was adjusted to allow a
nullptr CastSocket* argument, when a broadcast message comes from a
local peer instead of a remote sender.

Bug: b/169453993
Change-Id: Idbc698336e01022b8e7090498e882c66bb4922b0
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2433087
Commit-Queue: Yuri Wiitala <miu@chromium.org>
Reviewed-by: Brandon Tolsch <btolsch@chromium.org>
Reviewed-by: Yuri Wiitala <miu@chromium.org>
Reviewed-by: Jordan Bayles <jophba@chromium.org>