Add build flag for self-signed certs
This patch changes the standalone receiver and sender implementations to
use a new CAST_ALLOW_DEVELOPER_CERTIFICATE build flag/preprocessor
macro for gating usage of self-signed certificates. When false, the
cast_receiver app is disabled, and the cast_sender app will only connect
to receivers using a proper Google cert.

When the build flag is enabled, the cast sender can use any self-signed
certificate as long as the CA bit is set. The
cast receiver can either take a self-signed certificate and its private
key, or generate
both the private key and the certificate. The resulting private key and
certificate are written out to files, currently:
./generated_root_cast_receiver.(key|crt)

and can then be used with a
cast_sender to start a session.

Bug: b/169796278
Change-Id: I03675f85ac0b2bda76daf1bf11d9d7df064d0b8f
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/2443934
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Brandon Tolsch <btolsch@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>